+++ crates/anvil-polkadot/Cargo.toml
@@ -59,6 +59,7 @@ polkadot-sdk = { version = "2507.1.0", default-features = false, features = [
 	"substrate-frame-rpc-system",
 	"substrate-rpc-client",
 	"substrate-wasm-builder",
+	"pallet-revive-eth-rpc"
 ] }
 anvil-core.workspace = true
 anvil-server = { workspace = true, features = ["clap"] }
