+++ crates/anvil-polkadot/src/config.rs
@@ -10,7 +10,8 @@ use eyre::{Context, Result};
 use foundry_common::{duration_since_unix_epoch, sh_println};
 use polkadot_sdk::{
     sc_cli::{
-        self, CliConfiguration as SubstrateCliConfiguration, Cors, RPC_DEFAULT_MAX_CONNECTIONS,
+        self, CliConfiguration as SubstrateCliConfiguration, Cors, DatabaseParams,
+        ExecutionStrategiesParams, PruningParams, RPC_DEFAULT_MAX_CONNECTIONS,
         RPC_DEFAULT_MAX_REQUEST_SIZE_MB, RPC_DEFAULT_MAX_RESPONSE_SIZE_MB,
         RPC_DEFAULT_MAX_SUBS_PER_CONN, RPC_DEFAULT_MESSAGE_CAPACITY_PER_CONN,
     },
@@ -63,6 +64,7 @@ const BANNER: &str = r"
 pub struct SubstrateNodeConfig {
     shared_params: sc_cli::SharedParams,
     rpc_params: sc_cli::RpcParams,
+    import_params: sc_cli::ImportParams,
 }
 
 impl SubstrateNodeConfig {
@@ -98,7 +100,32 @@ impl SubstrateNodeConfig {
             rpc_cors: None,
         };
 
-        Self { shared_params, rpc_params }
+        let import_params = sc_cli::ImportParams {
+            pruning_params: PruningParams {
+                state_pruning: None,
+                blocks_pruning: sc_cli::DatabasePruningMode::Archive,
+            },
+            database_params: DatabaseParams {
+                database: Some(sc_cli::Database::ParityDb),
+                database_cache_size: None,
+            },
+            wasm_method: sc_cli::WasmExecutionMethod::Compiled,
+            wasmtime_instantiation_strategy:
+                sc_cli::WasmtimeInstantiationStrategy::PoolingCopyOnWrite,
+            wasm_runtime_overrides: None,
+            execution_strategies: ExecutionStrategiesParams {
+                execution_block_construction: None,
+                execution_import_block: None,
+                execution_offchain_worker: None,
+                execution_syncing: None,
+                execution_other: None,
+                execution: Some(sc_cli::ExecutionStrategy::Native),
+            },
+            trie_cache_size: 1024 * 1024 * 1024,
+            warm_up_trie_cache: None,
+        };
+
+        Self { shared_params, rpc_params, import_params }
     }
 }
 
@@ -108,7 +135,7 @@ impl SubstrateCliConfiguration for SubstrateNodeConfig {
     }
 
     fn import_params(&self) -> Option<&sc_cli::ImportParams> {
-        None
+        Some(&self.import_params)
     }
 
     fn network_params(&self) -> Option<&sc_cli::NetworkParams> {
