+++ crates/anvil-polkadot/src/substrate_node/service/mod.rs
@@ -1,3 +1,4 @@
+use codec::{Decode, Encode};
 use parking_lot::Mutex;
 use polkadot_sdk::{
     sc_basic_authorship, sc_consensus, sc_consensus_manual_seal,
@@ -14,9 +15,13 @@ use std::sync::Arc;
 use substrate_runtime::Block;
 
 use crate::AnvilNodeConfig;
-pub use client::{BackendWithOverlay, Client, StorageOverrides};
+pub use backend::{BackendWithOverlay, StorageOverrides};
+pub use client::Client;
 
+mod backend;
 mod client;
+mod executor;
+pub mod storage;
 
 pub type Backend = sc_service::TFullBackend<Block>;
